fix: comply with new biome check
